Add 100GbE tutorial documentation and material
This tutorial is written broken out into two parts. The first part is a simple
standalone use of the 100GbE YB. The second part integrates the RFDC with simple
packetization using a fifo and adding a sequence count. Packets are then caught
at a server with a simple python script to catch, unpack, and perform a simple
spectrometer on raw adc time samples.