red pitaya spectrometer fix
spectrometer python script hangs never advancing to produce plots. The reason
being that `sync_fft` was never toggling the fft's `start_frame_in` to run the
fft core resulting in a pulse on `start_frame_out`. Either something changed in
the underlying xilinx sysgen fft implementation requring `start_frame_in` to run
the fft, or the tutorial python script was not correctly updated.

Added a circuit to count valid cycles from the ADC for the duration of the FFT
frame and then pulse to start the next frame. Also updated for a 14-bit ADC.